requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
django-hashid-field==2.1.6
django-recaptcha==2.0.5
hashids==1.2.0
mysqlclient==1.4.4
pdfkit==0.6.1
Pillow==6.2.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
